ProjetoSQL/Teste1/requirements.txt
django==3.0
pyodbc==4.0.31
django-mssql-backend==2.8.1
django-grappelli==2.15.1
django-import-export

